refactor(story4.1): split constant class into more detailed classes
- refactor constant class into more specific classes
- change package location used by other classes
- add amount getter method in lottomachine for yield calculation
