+++ src/main/java/lotto/common/ErrorConstants.java
@@ -1,0 +1,5 @@
+package lotto.common;
+
+public class ErrorConstants {
+    public static final String ERROR_HEADER = "[ERROR] ";
+}
