+++ src/main/java/lotto/domain/WinningLotto.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
-import static lotto.common.Constants.LOTTO_NUMBER_MAX;
-import static lotto.common.Constants.LOTTO_NUMBER_MIN;
+import static lotto.common.LottoConstants.LOTTO_NUMBER_MAX;
+import static lotto.common.LottoConstants.LOTTO_NUMBER_MIN;
 
 import java.util.List;
 import lotto.common.LottoRank;
